+++ src/components/CHANGELOG.md
@@ -1,3 +1,7 @@
+##
+
+modify HeatList
+
 ## 1.1.1
 
 fix HeatList conflict